VIX-3697 Fixed a lot of the left-over code remaining from all the previous design iterations
