+++ src/Vixen.Modules/App/Props/Models/Line/LineModel.cs
@@ -43,6 +43,31 @@ namespace VixenModules.App.Props.Models.Line
             return nodePoints;
         }
 
+        /// <summary>
+        /// Rotates the NodePoints around the center of a 0,1 matrix.
+        /// </summary>
+        /// <param name="nodePoints"></param>
+        /// <param name="angleInDegrees"></param>
+        protected static void RotateNodePoints(List<NodePoint> nodePoints, int angleInDegrees)
+        {
+	        double centerX = .5;
+	        double centerY = .5;
+	        double angleInRadians = angleInDegrees * (Math.PI / 180);
+	        double cosTheta = Math.Cos(angleInRadians);
+	        double sinTheta = Math.Sin(angleInRadians);
+	        foreach (var nodePoint in nodePoints)
+	        {
+		        double x =
+			        cosTheta * (nodePoint.X - centerX) -
+			        sinTheta * (nodePoint.Y - centerY);
+		        double y =
+			        sinTheta * (nodePoint.X - centerX) +
+			        cosTheta * (nodePoint.Y - centerY);
+
+		        nodePoint.X = x + centerX;
+		        nodePoint.Y = y + centerY;
+	        }
+        }
 		/// <inheritdoc/>				
 		protected override IEnumerable<NodePoint> Get3DNodePoints()
 		{
